updated task route authorization so only admins can add/delete/get

+++ routes/tasksRoutes.js
@@ -14,23 +14,31 @@ const options = {
 
 
 router.get("/", async (req, res) => {
-  try {
-    const tasks = await db("tasks").select(
-      "id",
-      "name",
-      "volunteer_id",
-      "subject",
-      "description",
-      "date"
-    );
-    res.status(200).json(tasks);
-  } catch (error) {
-    res.status(500).json({ message: "there was an error with your request" });
+  const decoded = jwt.verify(req.headers.authorization, process.env.SECRET);
+  if(decoded.role === 'admin') {
+    try {
+      const tasks = await db("tasks").select(
+        "id",
+        "name",
+        "volunteer_id",
+        "subject",
+        "description",
+        "date"
+      );
+      res.status(200).json(tasks);
+    } catch (error) {
+      res.status(500).json({ message: "there was an error with your request" });
+    }
+  } else {
+    res.status(401).json({message: "not authorized"})
   }
+
 });
 
 router.post("/", async (req, res) => {
-  const { name, volunteer, subject, description, date } = req.body;
+  const decoded = jwt.verify(req.headers.authorization, process.env.SECRET);
+  if(decoded.role === 'admin') {
+    const { name, volunteer, subject, description, date } = req.body;
   const schema = Joi.object({
     name: Joi.string().min(3).max(60).required(),
     volunteer: Joi.number().required(),
@@ -56,42 +64,52 @@ router.post("/", async (req, res) => {
     });
     res.status(400).json({ error: error });
   } else {
-    try {
-      await db("tasks").insert({
-        name: name,
-        volunteer_id: volunteer,
-        subject: subject,
-        description: description,
-        date: date,
-      });
-      res.status(201).end();
-    } catch (error) {
-      console.log(error);
-      res.status(500).json({ message: "could not add task" });
-    }
+      try {
+        await db("tasks").insert({
+          name: name,
+          volunteer_id: volunteer,
+          subject: subject,
+          description: description,
+          date: date,
+        });
+        res.status(201).end();
+      } catch (error) {
+        console.log(error);
+        res.status(500).json({ message: "could not add task" });
+      }
+    } 
+  } else {
+    res.status(401).json({message: "not authorized"})
   }
 });
 
 
 router.delete("/:id", async (req, res) => {
-  const { id } = req.params;
-  try {
-    const task = await db("tasks").where({ id });
-    if (task.length === 1) {
-      try {
-        const taskToDelete = task[0].id;
-        await db("tasks").where({ id: taskToDelete }).delete();
-        res.status(200).json({ message: "task deleted" });
-      } catch (error) {
-        console.log(error);
-        res.status(500).json({ message: "could not process your request" });
+  const decoded = jwt.verify(req.headers.authorization, process.env.SECRET);
+  if(decoded.role === 'admin') {
+    const { id } = req.params;
+    try {
+      const task = await db("tasks").where({ id });
+      if (task.length === 1) {
+        try {
+          const taskToDelete = task[0].id;
+          await db("tasks").where({ id: taskToDelete }).delete();
+          res.status(200).json({ message: "task deleted" });
+        } catch (error) {
+          console.log(error);
+          res.status(500).json({ message: "could not process your request" });
+        }
+      } else {
+        res.status(400).json({ message: "task not found" });
       }
-    } else {
-      res.status(400).json({ message: "task not found" });
+    } catch (error) {
+      console.log(error);
+      res.status(500).json({ message: "could not process your request" });
     }
-  } catch (error) {
-    console.log(error);
-    res.status(500).json({ message: "could not process your request" });
+  } else {
+    res.status(401).json({message: "not authorized"})
   }
+  
+
 });
 module.exports = router;
